+++ Source/DataTransfer.Legacy.MassImport/Data/Objects.cs
@@ -157,7 +157,7 @@ namespace Relativity.MassImport.Data
 			sql.Add(new InlineSqlQuery($"DECLARE @now DATETIME = GETUTCDATE()"));
 
 			sql.Add(this.ImportSql.PopulateAssociatedPartTable(this._tableNames, field.ArtifactID, associatedObjectTable, field.GetColumnName(), idFieldColumnName));
-			sql.Add(AssociatedObjectsValidationSql.ValidateAssociatedObjectsForSingleObjectField(this._tableNames, field, associatedArtifactTypeId));
+			sql.Add(AssociatedObjectsValidationSql.ValidateAssociatedObjectsForSingleObjectField(this._tableNames, field, associatedArtifactTypeId, associatedObjectTable));
 
 			if (associatedArtifactTypeId == (int) ArtifactType.Document)
 			{