+++ Source/DataTransfer.Legacy.MassImport/RelEyeTelemetry/Events/EventGeneralStatistics.cs
@@ -1,0 +1,13 @@
+﻿// <copyright file="EventGeneralStatistics.cs" company="Relativity ODA LLC">
+// © Relativity All Rights Reserved.
+// </copyright>
+
+namespace DataTransfer.Legacy.MassImport.RelEyeTelemetry.Events
+{
+	/// <inheritdoc/>
+	public class EventGeneralStatistics : EventBase
+	{
+		/// <inheritdoc/>
+		public override string EventName => Constants.EventName.GeneralStatistics;
+	}
+}