+++ Source/DataTransfer.Legacy.Services.Tests/HealthCheckServiceTests.cs
@@ -1,92 +1,46 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Threading.Tasks;
-using Castle.MicroKernel.Registration;
-using Castle.Windsor;
+﻿using System.Threading.Tasks;
 using Moq;
 using NUnit.Framework;
-using Relativity.API;
 using Relativity.Core;
 using Relativity.DataTransfer.Legacy.SDK.ImportExport.V1;
 using Relativity.DataTransfer.Legacy.SDK.ImportExport.V1.Models;
-using Relativity.DataTransfer.Legacy.Services.Helpers;
-using Relativity.DataTransfer.Legacy.Services.Interceptors;
-using Relativity.DataTransfer.Legacy.Services.Metrics;
-using Relativity.DataTransfer.Legacy.Services.Observability;
-using Relativity.Services.Interfaces.LibraryApplication;
 using TddEbook.TddToolkit;
 
 namespace Relativity.DataTransfer.Legacy.Services.Tests
 {
-    [TestFixture]
-	public class HealthCheckServiceTests
+	[TestFixture]
+	public class HealthCheckServiceTests : ServicesBaseTests
 	{
-		private Mock<IServiceContextFactory> _serviceContextFactoryMock;
-		private Mock<ICommunicationModeStorage> _communicationModeStorageMock;
-		private Mock<IRelativityPermissionHelper> _relativityPermissionHelperMock;
-		private Mock<IAPILog> _loggerMock;
-		private Mock<IMetricsContext> _metricsMock;
-		private Mock<IMetricsPublisher> _metricsPublisherMock;
-		private Mock<ILibraryApplicationManager> _libraryApplicationManager;
 		private IHealthCheckService _uut;
 
 		[SetUp]
 		public void SetUp()
 		{
-			_serviceContextFactoryMock = new Mock<IServiceContextFactory>();
-			_communicationModeStorageMock = new Mock<ICommunicationModeStorage>();
-			_relativityPermissionHelperMock = new Mock<IRelativityPermissionHelper>();
-			_loggerMock = new Mock<IAPILog>();
-			_metricsMock = new Mock<IMetricsContext>();
-			_metricsPublisherMock = new Mock<IMetricsPublisher>();
-			_libraryApplicationManager = new Mock<ILibraryApplicationManager>();
-
-			var container = new WindsorContainer();
-			container.Register(Component.For<IAPILog>().Instance(_loggerMock.Object));
-			container.Register(Component.For<IServiceContextFactory>().Instance(_serviceContextFactoryMock.Object));
-			container.Register(Component.For<ICommunicationModeStorage>().Instance(_communicationModeStorageMock.Object));
-			container.Register(Component.For<IRelativityPermissionHelper>().Instance(_relativityPermissionHelperMock.Object));
-			container.Register(Component.For<ToggleCheckInterceptor>());
-			container.Register(Component.For<PermissionCheckInterceptor>());
-			container.Register(Component.For<LogInterceptor>());
-			container.Register(Component.For<MetricsInterceptor>());
-			container.Register(Component.For<UnhandledExceptionInterceptor>());
-			container.Register(Component.For<DistributedTracingInterceptor>());
-			container.Register(Component.For<IMetricsContext>().Instance(_metricsMock.Object));
-			container.Register(Component.For<IMetricsPublisher>().Instance(_metricsPublisherMock.Object));
-			container.Register(Component.For<IEnumerable<IMetricsPublisher>>()
-				.Instance(new[] {_metricsPublisherMock.Object}));
-			container.Register(Component.For<Func<IMetricsContext>>().UsingFactoryMethod(x =>
-				new Func<IMetricsContext>(container.Resolve<IMetricsContext>)));
-			container.Register(Component.For<ILibraryApplicationManager>().Instance(_libraryApplicationManager.Object));
-			container.Register(Component.For<IHealthCheckService>().ImplementedBy<HealthCheckService>());
-			container.Register(Component.For<ITraceGenerator>().ImplementedBy<TraceGenerator>());
-
-			_uut = container.Resolve<IHealthCheckService>();
+			_uut = Container.Resolve<IHealthCheckService>();
 		}
 
 
 		[Test]
 		public void ShouldExecuteWhenCommunicationModeTurnsKeplerServicesOff()
 		{
-			_communicationModeStorageMock.Setup(x => x.TryGetModeAsync())
+			CommunicationModeStorageMock.Setup(x => x.TryGetModeAsync())
 				.Returns(Task.FromResult((true, IAPICommunicationMode.ForceWebAPI)));
 
 			_uut.HealthCheckAsync();
-			
+
 		}
 
 		[Test]
 		public void ShouldExecuteRegardlessPermissions()
 		{
-			_communicationModeStorageMock.Setup(x => x.TryGetModeAsync())
+			CommunicationModeStorageMock.Setup(x => x.TryGetModeAsync())
 				.Returns(Task.FromResult((true, Any.OtherThan(IAPICommunicationMode.ForceWebAPI))));
 			var baseServiceContext = Any.InstanceOf<BaseServiceContext>();
-			_serviceContextFactoryMock.Setup(x => x.GetBaseServiceContext(It.IsAny<int>()))
+			ServiceContextFactoryMock.Setup(x => x.GetBaseServiceContext(It.IsAny<int>()))
 				.Returns(baseServiceContext);
-			_relativityPermissionHelperMock.Setup(x =>
+			RelativityPermissionHelperMock.Setup(x =>
 				x.HasAdminOperationPermission(baseServiceContext, Permission.AllowDesktopClientImport)).Returns(false);
-			_relativityPermissionHelperMock.Setup(x =>
+			RelativityPermissionHelperMock.Setup(x =>
 				x.HasAdminOperationPermission(baseServiceContext, Permission.AllowDesktopClientExport)).Returns(false);
 
 			_uut.HealthCheckAsync();