+++ Source/DataTransfer.Legacy.Services/Installer/ServicesInstaller.cs
@@ -11,10 +11,13 @@ using Relativity.DataTransfer.Legacy.Services.Metrics;
 using Relativity.DataTransfer.Legacy.Services.Observability;
 using Relativity.DataTransfer.Legacy.Services.SQL;
 using Relativity.Telemetry.APM;
+using DataTransfer.Legacy.MassImport.Data;
+using DataTransfer.Legacy.MassImport.RelEyeTelemetry;
+using DataTransfer.Legacy.MassImport.RelEyeTelemetry.MetricsEventsBuilders;
 
 namespace Relativity.DataTransfer.Legacy.Services.Installer
 {
-    public class ServicesInstaller : IWindsorInstaller
+	public class ServicesInstaller : IWindsorInstaller
 	{
 		public void Install(IWindsorContainer container, IConfigurationStore store)
 		{
@@ -29,6 +32,7 @@ namespace Relativity.DataTransfer.Legacy.Services.Installer
 			// interceptors registration
 			container.Register(Component.For<LogInterceptor>().LifestyleTransient());
 			container.Register(Component.For<MetricsInterceptor>().LifestyleTransient());
+			container.Register(Component.For<DistributedTracingInterceptor>().LifestyleTransient());
 			container.Register(Component.For<PermissionCheckInterceptor>().LifestyleTransient());
 			container.Register(Component.For<UnhandledExceptionInterceptor>().LifestyleTransient());
 			container.Register(Component.For<ToggleCheckInterceptor>().LifestyleTransient());
@@ -51,6 +55,10 @@ namespace Relativity.DataTransfer.Legacy.Services.Installer
 			container.Register(Component.For<ISqlRetryPolicy>().ImplementedBy<SqlRetryPolicy>().LifestyleTransient());
 			container.Register(Component.For<ITraceGenerator>().ImplementedBy<TraceGenerator>().LifestyleTransient());
 			container.Register(Component.For<RetryPolicyFactory>());
+			container.Register(Component.For<ITelemetryPublisher>().ImplementedBy<ApmTelemetryPublisher>().LifestyleTransient());
+			container.Register(Component.For<IRelEyeMetricsService>().ImplementedBy<RelEyeMetricsService>().LifestyleTransient());
+			container.Register(Component.For<IEventsBuilder>().ImplementedBy<EventsBuilder>().LifestyleTransient());
+			StorageAccessProvider.InitializeStorageAccess(container);
 		}
 	}
 }